+++ packages/typespec-rust/src/codemodel/types.ts
@@ -258,7 +258,7 @@ export interface PageIterator extends External {
   kind: 'pageIterator';
 
   /** the model containing the page of items */
-  type: Response<Payload<Model>>;
+  type: Response<Model, Exclude<ResponseFormat, 'NoFormat'>>;
 }
 
 /** Pager is a Pager<T> from azure_core */
@@ -266,11 +266,14 @@ export interface Pager extends External {
   kind: 'pager';
 
   /** the model containing the page of items */
-  type: Payload<Model>;
+  type: Response<Model, Exclude<ResponseFormat, 'NoFormat'>>;
 }
 
+/** PayloadFormat indicates the wire format for request bodies */
+export type PayloadFormat = 'json' | 'xml';
+
 /**
- * Payload<T> is used for operations that send/receive a typed payload.
+ * Payload<T> is used for operations that send a typed payload.
  * it's a grouping of the payload type and its wire format but does not
  * actually exist as a type (i.e. there's no Payload type in emitted code).
  */
@@ -283,7 +286,7 @@ export interface Payload<T extends WireType = WireType> {
   type: T;
 
   /** the wire format of the request body */
-  format: BodyFormat;
+  format: PayloadFormat;
 }
 
 /**
@@ -325,15 +328,21 @@ export interface RequestContent<T extends RequestContentTypes = RequestContentTy
   content: T;
 }
 
-/** ResponseTypes defines the type constraint when creating a Response<T> */
-export type ResponseTypes = MarkerType | Payload | Unit;
+/** ResponseFormat is the format of the response body */
+export type ResponseFormat = 'JsonFormat' | 'NoFormat' | 'XmlFormat';
 
-/** Response is a Rust Response<T> from azure_core */
-export interface Response<T extends ResponseTypes = ResponseTypes> extends External {
+/** ResponseTypes defines the type constraint when creating a Response<T> */
+export type ResponseTypes = MarkerType | Unit | WireType;
+
+/** Response is a Rust Response<T, Format> from azure_core */
+export interface Response<T extends ResponseTypes = ResponseTypes, Format extends ResponseFormat = ResponseFormat> extends External {
   kind: 'response';
 
   /** the type of content sent in the response */
   content: T;
+
+  /** the wire format of the response body */
+  format: Format;
 }
 
 /** ResultTypes defines the type constraint when creating a Result<T> */
@@ -368,9 +377,6 @@ export interface Scalar {
   /** indicates that the value is encoded/decoded as a string */
   stringEncoding: boolean;
 }
-
-/** BodyFormat indicates the wire format for request and response bodies */
-export type BodyFormat = 'json' | 'xml';
 
 /** Slice is a Rust slice i.e. [T] */
 export interface Slice {
@@ -697,7 +703,7 @@ export class Option implements Option {
 }
 
 export class PageIterator extends External implements PageIterator {
-  constructor(crate: Crate, type: Response<Payload<Model>>) {
+  constructor(crate: Crate, type: Response<Model, Exclude<ResponseFormat, 'NoFormat'>>) {
     super(crate, 'PageIterator', 'azure_core::http');
     this.kind = 'pageIterator';
     this.type = type;
@@ -705,7 +711,7 @@ export class PageIterator extends External implements PageIterator {
 }
 
 export class Pager extends External implements Pager {
-  constructor(crate: Crate, type: Payload<Model>) {
+  constructor(crate: Crate, type: Response<Model, Exclude<ResponseFormat, 'NoFormat'>>) {
     super(crate, 'Pager', 'azure_core::http');
     this.kind = 'pager';
     this.type = type;
@@ -713,7 +719,7 @@ export class Pager extends External implements Pager {
 }
 
 export class Payload<T> implements Payload<T> {
-  constructor(type: T, format: BodyFormat) {
+  constructor(type: T, format: PayloadFormat) {
     this.kind = 'payload';
     this.type = type;
     this.format = format;
@@ -742,11 +748,12 @@ export class RequestContent<T> extends External implements RequestContent<T> {
   }
 }
 
-export class Response<T> extends External implements Response<T> {
-  constructor(crate: Crate, content: T) {
+export class Response<T, Format> extends External implements Response<T, Format> {
+  constructor(crate: Crate, content: T, format: Format) {
     super(crate, 'Response', 'azure_core::http');
     this.kind = 'response';
     this.content = content;
+    this.format = format;
   }
 }